Add, improved exporting, still should be refactored
